feat: separate Spinner and SpinnerIcon components
icon can be used in other loading states such as toast
