+++ react/src/Spinner/index.ts
@@ -1,1 +1,2 @@
 export * from './Spinner'
+export * from './SpinnerIcon'
